# adapter/input/jakarta-controller/src/test/resources/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Real State API",
    "description" : "This API helps manage all properties that are for sale, rent, generate contracts and much more",
    "contact" : {
      "name" : "Real Estate",
      "url" : "https://realestate.com.br",
      "email" : "contato@realestate.com.br"
    },
    "version" : "1"
  },
  "paths" : {
    "/v1/owner" : {
      "post" : {
        "summary" : "Create an owner",
        "operationId" : "createOwner",
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/OwnerRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Return the created owner",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          },
          "409" : {
            "description" : "There is a owner with the same information"
          },
          "422" : {
            "description" : "Unprocessable Entity",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          }
        }
      }
    },
    "/v1/property" : {
      "get" : {
        "summary" : "Get all properties",
        "operationId" : "findAll",
        "parameters" : [ {
          "name" : "_limit",
          "in" : "query",
          "schema" : {
            "maximum" : 50,
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        }, {
          "name" : "_offset",
          "in" : "query",
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "_sort",
          "in" : "query",
          "schema" : {
            "type" : "string",
            "default" : "ASC"
          }
        }, {
          "name" : "propertyKind",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "country",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "state",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "city",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "neighborhood",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "schema" : {
            "minimum" : 10,
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "rooms",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "isRent",
          "in" : "query",
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "isSale",
          "in" : "query",
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "isFurnished",
          "in" : "query",
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "totalPrice",
          "in" : "query",
          "schema" : {
            "type" : "number",
            "description" : "This field is mandatory if isRent equals True"
          }
        }, {
          "name" : "garage",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Return a list of properties",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/PropertyResponseDto"
                  }
                }
              }
            }
          },
          "204" : {
            "description" : "empty list"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "OwnerRequestDto" : {
        "required" : [ "city", "country", "document", "document type", "email", "name", "neighborhood", "number", "phoneNumber", "state", "street_name", "zip_code" ],
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "document" : {
            "type" : "string"
          },
          "document type" : {
            "type" : "string",
            "enum" : [ "CPF", "CNPJ" ]
          },
          "country" : {
            "type" : "string"
          },
          "state" : {
            "maxLength" : 2,
            "minLength" : 2,
            "type" : "string",
            "description" : "We expect an abbreviation of the state"
          },
          "city" : {
            "type" : "string"
          },
          "neighborhood" : {
            "type" : "string"
          },
          "zip_code" : {
            "type" : "string"
          },
          "street_name" : {
            "type" : "string"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "complement" : {
            "type" : "string"
          },
          "phoneNumber" : {
            "type" : "string"
          }
        }
      },
      "PropertyResponseDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "propertyKind" : {
            "type" : "string"
          },
          "country" : {
            "type" : "string"
          },
          "state" : {
            "type" : "string"
          },
          "city" : {
            "type" : "string"
          },
          "neighborhood" : {
            "type" : "string"
          },
          "zipCode" : {
            "type" : "string"
          },
          "streetName" : {
            "type" : "string"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "complement" : {
            "type" : "string"
          },
          "landSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "buildingArea" : {
            "type" : "integer",
            "format" : "int32"
          },
          "rooms" : {
            "type" : "integer",
            "format" : "int32"
          },
          "isRent" : {
            "type" : "boolean"
          },
          "isSale" : {
            "type" : "boolean"
          },
          "isFurnished" : {
            "type" : "boolean"
          },
          "rentPrice" : {
            "type" : "number"
          },
          "salePrice" : {
            "type" : "number"
          },
          "taxes" : {
            "type" : "number"
          },
          "isCondominium" : {
            "type" : "boolean"
          },
          "condominiumValue" : {
            "type" : "number"
          },
          "garage" : {
            "type" : "integer",
            "format" : "int32"
          },
          "description" : {
            "type" : "string"
          },
          "propertyItems" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "condominiumItems" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "images" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          }
        }
      }
    }
  }
}